Reorganize programming question files
