+++ client/app/bundles/course/assessment/question/programming/components/OnlineEditor/ExistingPackageFile.jsx
@@ -6,8 +6,16 @@ import {
 } from 'material-ui/Table';
 import RaisedButton from 'material-ui/RaisedButton';
 import { grey100, grey300, white } from 'material-ui/styles/colors';
-import styles from './../containers/OnlineEditor/OnlineEditorView.scss';
-import { formatBytes } from './../reducers/utils';
+import styles from './OnlineEditorView.scss';
+
+function formatBytes(bytes, decimals) {
+  if (bytes === 0) return '0 Byte';
+  const k = 1000; // or 1024 for binary
+  const dm = decimals || 3;
+  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return `${parseFloat((bytes / (k ** i)).toFixed(dm))} ${sizes[i]}`;
+}
 
 function ExistingPackageFile(props) {
   const { filename, fileType, filesize, toDelete, deleteExistingPackageFile, isLoading, isLast } = props;
