+++ client/app/bundles/course/assessment/question/programming/components/OnlineEditor/Java/OnlineEditorJavaView.jsx
@@ -2,28 +2,31 @@ import Immutable from 'immutable';
 
 import React from 'react';
 import PropTypes from 'prop-types';
+import AceEditor from 'react-ace';
 import { injectIntl, FormattedMessage, intlShape } from 'react-intl';
 import { Card, CardHeader, CardText } from 'material-ui/Card';
 import FlatButton from 'material-ui/FlatButton';
+import Toggle from 'material-ui/Toggle';
 import {
   Table, TableBody, TableFooter, TableHeader, TableHeaderColumn, TableRow, TableRowColumn,
 } from 'material-ui/Table';
 import transitions from 'material-ui/styles/transitions';
 
-import 'brace/mode/c_cpp';
+import 'brace/mode/java';
 import 'brace/theme/monokai';
 
-import styles from './../OnlineEditorView.scss';
-import translations from './../OnlineEditorView.intl';
-import cppTranslations from './OnlineEditorCppView.intl';
-import { ExistingPackageFile, NewPackageFile, TestCase, EditorCard } from './../OnlineEditorBase';
+import styles from '../OnlineEditorView.scss';
+import translations from '../OnlineEditorView.intl';
+import javaTranslations from './OnlineEditorJavaView.intl';
+import { ExistingPackageFile, NewPackageFile, TestCase, EditorCard } from '../OnlineEditorBase';
 
 const MAX_TEST_CASES = 99;
 
 const propTypes = {
   data: PropTypes.instanceOf(Immutable.Map).isRequired,
-  dataFiles: PropTypes.instanceOf(Immutable.Map).isRequired,
+  testData: PropTypes.instanceOf(Immutable.Map).isRequired,
   actions: PropTypes.shape({
+    toggleSubmitAsFile: PropTypes.func.isRequired,
     updateCodeBlock: PropTypes.func.isRequired,
     createTestCase: PropTypes.func.isRequired,
     updateTestCase: PropTypes.func.isRequired,
@@ -34,6 +37,7 @@ const propTypes = {
   }),
   isLoading: PropTypes.bool.isRequired,
   autograded: PropTypes.bool.isRequired,
+  hasSubmissions: PropTypes.bool.isRequired,
   intl: intlShape.isRequired,
 };
 
@@ -41,7 +45,11 @@ const contextTypes = {
   muiTheme: PropTypes.object.isRequired,
 };
 
-class OnlineEditorCppView extends React.Component {
+class OnlineEditorJavaView extends React.Component {
+  static getTestInputName(type, field) {
+    return `question_programming[test_cases][${type}][][${field}]`;
+  }
+
   testCaseCreateHandler(type) {
     return (e) => {
       e.preventDefault();
@@ -69,7 +77,7 @@ class OnlineEditorCppView extends React.Component {
             filename,
             fileType,
             filesize: fileData.get('size'),
-            toDelete: this.props.dataFiles.get('to_delete').has(filename),
+            toDelete: this.props.testData.getIn([fileType, 'to_delete']).has(filename),
             deleteExistingPackageFile: this.props.actions.deleteExistingPackageFile,
             isLoading: this.props.isLoading,
             isLast: numFiles === index + 1,
@@ -100,7 +108,7 @@ class OnlineEditorCppView extends React.Component {
               </TableRow>
             </TableHeader>
             <TableBody displayRowCheckbox={false}>
-              {this.props.data.get('data_files').map(renderDataFile)}
+              {this.props.data.get(fileType).map(renderDataFile)}
             </TableBody>
           </Table>
         </CardText>
@@ -118,7 +126,7 @@ class OnlineEditorCppView extends React.Component {
             index,
             fileType,
             filename: fileData.get('filename'),
-            showDeleteButton: this.props.dataFiles.get('key') !== key,
+            showDeleteButton: this.props.testData.getIn([fileType, 'key']) !== key,
             isLoading: this.props.isLoading,
             updateNewPackageFile: this.props.actions.updateNewPackageFile,
             deleteNewPackageFile: this.props.actions.deleteNewPackageFile,
@@ -127,7 +135,7 @@ class OnlineEditorCppView extends React.Component {
         />
       );
     };
-    const newPackageFilesRows = this.props.dataFiles.get('new').map(renderNewFile);
+    const newPackageFilesRows = this.props.testData.getIn([fileType, 'new']).map(renderNewFile);
 
     return (
       <Card initiallyExpanded>
@@ -171,12 +179,52 @@ class OnlineEditorCppView extends React.Component {
           expression,
           expected,
           hint,
-          enableInlineCodeEditor: false,
-          showCodeEditor: test.get('showCodeEditor') || false,
+          enableInlineCodeEditor: true,
+          showCodeEditor: test.get('show_code_editor') || false,
           intl: this.props.intl,
         }}
       />
     ));
+
+    const editors = [...testCases.get(type).entries()].map(([index, test]) => {
+      const showCodeEditor = test.get('show_code_editor') || false;
+      const inlineCode = test.get('inline_code') || '';
+      let editorStyle = { display: 'none' };
+      if (showCodeEditor) {
+        editorStyle = {};
+      }
+      return (
+        <TableRow id={index} style={editorStyle}>
+          <TableRowColumn colSpan="6" style={{ textAlign: 'center', paddingLeft: 0, paddingRight: 0 }}>
+            <textarea
+              name={OnlineEditorJavaView.getTestInputName(type, 'inline_code')}
+              value={test.get('inline_code')}
+              style={{ display: 'none' }}
+              readOnly="true"
+            />
+            <AceEditor
+              mode="java"
+              theme="monokai"
+              width="100%"
+              minLines={4}
+              maxLines={Math.max(20, inlineCode.split(/\r\n|\r|\n/).length)}
+              name={OnlineEditorJavaView.getTestInputName(type, 'inline_code')}
+              value={test.get('inline_code')}
+              onChange={e => this.props.actions.updateTestCase(type, index, 'inline_code', e)}
+              editorProps={{ $blockScrolling: true }}
+              setOptions={{ useSoftTabs: true, readOnly: this.props.isLoading }}
+            />
+          </TableRowColumn>
+        </TableRow>
+      );
+    });
+
+    // To interweave the two arrays
+    const testCaseRows = [];
+    rows.forEach((val, index) => {
+      testCaseRows.push(val);
+      testCaseRows.push(editors[index]);
+    });
 
     return (
       <Card initiallyExpanded>
@@ -195,14 +243,15 @@ class OnlineEditorCppView extends React.Component {
                 <TableHeaderColumn>{expression}</TableHeaderColumn>
                 <TableHeaderColumn>{expected}</TableHeaderColumn>
                 <TableHeaderColumn>{hint}</TableHeaderColumn>
+                <TableHeaderColumn />
               </TableRow>
             </TableHeader>
             <TableBody displayRowCheckbox={false}>
-              {rows}
+              {testCaseRows}
             </TableBody>
             <TableFooter adjustForCheckbox={false}>
               <TableRow>
-                <TableRowColumn colSpan="5" style={{ textAlign: 'center' }}>
+                <TableRowColumn colSpan="6" style={{ textAlign: 'center' }}>
                   <FlatButton
                     label={this.props.intl.formatMessage(translations.addNewTestButton)}
                     icon={<i className="fa fa-plus" />}
@@ -224,7 +273,7 @@ class OnlineEditorCppView extends React.Component {
       <EditorCard
         {...{
           updateCodeBlock: this.props.actions.updateCodeBlock,
-          mode: 'c_cpp',
+          mode: 'java',
           field,
           value,
           header,
@@ -237,6 +286,7 @@ class OnlineEditorCppView extends React.Component {
 
   renderAutogradedFields() {
     const { intl, data } = this.props;
+    const submitAsFile = this.props.data.get('submit_as_file');
     const testCases = data.get('test_cases');
     const testCaseError = data.getIn(['test_cases', 'error']);
     const errorTextElement = testCaseError && (
@@ -257,23 +307,41 @@ class OnlineEditorCppView extends React.Component {
       <React.Fragment>
         <div style={{ marginBottom: '1em' }}>
           {
-            this.renderEditorCard(
-              intl.formatMessage(translations.solutionTitle),
-              intl.formatMessage(translations.solutionSubtitle),
-              'solution'
-            )
+            submitAsFile ?
+              <div style={{ marginBottom: '20px' }}>
+                <h3>{this.props.intl.formatMessage(javaTranslations.solutionFilesHeader)}</h3>
+                {
+                  this.renderExistingPackageFiles(
+                    'solution_files',
+                    this.props.intl.formatMessage(javaTranslations.currentSolutionFilesHeader)
+                    )
+                }
+                {
+                  this.renderNewPackageFiles(
+                    'solution_files',
+                    this.props.intl.formatMessage(javaTranslations.newSolutionFilesHeader),
+                    intl.formatMessage(javaTranslations.addSolutionFileButton)
+                  )
+                }
+              </div>
+              :
+              this.renderEditorCard(
+                intl.formatMessage(translations.solutionTitle),
+                intl.formatMessage(translations.solutionSubtitle),
+                'solution'
+              )
           }
           {
             this.renderEditorCard(
               intl.formatMessage(translations.prependTitle),
-              intl.formatMessage(translations.prependSubtitle),
+              intl.formatMessage(javaTranslations.prependSubtitle),
               'prepend'
             )
           }
           {
             this.renderEditorCard(
               intl.formatMessage(translations.appendTitle),
-              intl.formatMessage(cppTranslations.appendSubtitle),
+              intl.formatMessage(javaTranslations.appendSubtitle),
               'append'
             )
           }
@@ -295,27 +363,51 @@ class OnlineEditorCppView extends React.Component {
         <h3>{ intl.formatMessage(translations.testCasesHeader) }</h3>
         <div style={{ marginBottom: '0.5em' }}>
           <FormattedMessage
-            id="course.assessment.question.programming.onlineEditorCppView.testCasesDescription"
+            id="course.assessment.question.programming.onlineEditorJavaView.testCasesDescription"
             defaultMessage={
               '{note}: The expression in the {expression} column will be compared with the ' +
-              'expression in the {expected} column using {expect_star} assertions from the ' +
-              '{googletest}. Floating point numbers are formatted with {tostring}.'
+              'expression in the {expected} column using the {expectEquals} method as described ' +
+              'in the append.'
             }
             values={{
               note: <b>{intl.formatMessage(translations.testCaseDescriptionNote)}</b>,
               expression: <b>{intl.formatMessage(translations.expressionHeader)}</b>,
               expected: <b>{intl.formatMessage(translations.expectedHeader)}</b>,
-              expect_star: <code>EXPECT_*</code>,
-              googletest: (
-                <a href="https://github.com/google/googletest">
-                  {intl.formatMessage(translations.testCaseDescriptionGoogleTest)}
-                </a>
-              ),
-              tostring: (
-                <code>
-                  <a href="http://en.cppreference.com/w/cpp/string/basic_string/to_string">std::to_string</a>
-                </code>
-              ),
+              expectEquals: <code>{intl.formatMessage(javaTranslations.expectEquals)}</code>,
+            }}
+          />
+        </div>
+        <div style={{ marginBottom: '0.5em' }}>
+          <FormattedMessage
+            id="course.assessment.question.programming.onlineEditorJavaView.testCasesCodeDescription"
+            defaultMessage={
+              '{editor}: Clicking {code} will toggle a code editor for you to write code into each ' +
+              'test case. This allows you to initialize variables and call functions for each test. ' +
+              'For example: {codeExample}' +
+              '{codeExampleExpression} and {codeExampleExpected} can then be input into {expression} and ' +
+              '{expected} respectively.'
+            }
+            values={{
+              expression: <b>{intl.formatMessage(translations.expressionHeader)}</b>,
+              expected: <b>{intl.formatMessage(translations.expectedHeader)}</b>,
+              code: <b>{intl.formatMessage(javaTranslations.testCaseDescriptionCode)}</b>,
+              editor: <b>{intl.formatMessage(javaTranslations.testCaseDescriptionEditor)}</b>,
+              /* eslint-disable react/jsx-indent */
+              codeExample:
+                <pre style={{ marginTop: '0.5em', paddingTop: '3px', paddingBottom: '3px' }}>
+                  <p style={{ marginBottom: 0 }}>
+                    {'int array [] = {0,0,0}; // Initialize variables'}
+                  </p>
+                  <p style={{ marginBottom: 0 }}>addOneToArray(array); // Make function calls</p>
+                  <p style={{ marginBottom: 0 }}>int expected [] = {'{'}1,1,1{'}'}; // Make function calls</p>
+                  <p style={{ marginBottom: 0 }}>
+                    {'setAttribute("expression", "addOneToArray([0,0,0])");'}
+                    {' // Override the default expression displayed'}
+                  </p>
+                </pre>,
+              /* eslint-enable react/jsx-indent */
+              codeExampleExpected: <code>expected</code>,
+              codeExampleExpression: <code>array</code>,
             }}
           />
         </div>
@@ -337,16 +429,69 @@ class OnlineEditorCppView extends React.Component {
   }
 
   render() {
-    const { intl, autograded } = this.props;
-
+    const { intl, autograded, isLoading, hasSubmissions } = this.props;
+    const submitAsFile = this.props.data.get('submit_as_file');
+    let toggleLabel = intl.formatMessage(javaTranslations.submitAsFileToggle);
+    if (hasSubmissions) {
+      toggleLabel = intl.formatMessage(javaTranslations.submitAsFileToggleDisabled);
+    }
     return (
-      <div id="cpp-online-editor">
+      <div id="java-online-editor">
         {
-          this.renderEditorCard(
-            intl.formatMessage(translations.submissionTitle),
-            null,
-            'submission'
-          )
+          autograded &&
+          <div className={styles.submitAsFileToggle}>
+            <Toggle
+              label={toggleLabel}
+              labelPosition="right"
+              toggled={submitAsFile}
+              onToggle={(e) => {
+                if (hasSubmissions) return;
+                this.props.actions.toggleSubmitAsFile(e.target.checked);
+              }}
+              readOnly={hasSubmissions}
+              disabled={isLoading}
+              style={{ margin: '1em 0', paddingTop: 10 }}
+            />
+            <input hidden name="question_programming[submit_as_file]" value={submitAsFile} />
+            <FormattedMessage
+              id="course.assessment.question.programming.onlineEditorJavaView.fileSubmissionDescription"
+              defaultMessage={
+                '{file_submission}: Toggling this option on will allow you to upload java class files to be ' +
+                'compiled individually, and allows you to test (individual/multiple) java classes. ' +
+                'Toggled off, you will input code as templates, which will be used for you to test ' +
+                'java functions. Note that you will need to upload either a submission or solution ' +
+                'file at the very least for the compiler to compile the files correctly.'
+              }
+              values={{
+                file_submission: <b>{intl.formatMessage(javaTranslations.fileSubmissionDescriptionNote)}</b>,
+              }}
+            />
+          </div>
+        }
+        {
+          submitAsFile ?
+            <React.Fragment>
+              <h3>{this.props.intl.formatMessage(javaTranslations.submissionFilesHeader)}</h3>
+              {
+                this.renderExistingPackageFiles(
+                  'submission_files',
+                  this.props.intl.formatMessage(javaTranslations.currentSubmissionFilesHeader)
+                )
+              }
+              {
+                this.renderNewPackageFiles(
+                  'submission_files',
+                  this.props.intl.formatMessage(javaTranslations.newSubmissionFilesHeader),
+                  intl.formatMessage(javaTranslations.addSubmissionFileButton)
+                )
+              }
+            </React.Fragment>
+            :
+            this.renderEditorCard(
+              intl.formatMessage(translations.submissionTitle),
+              null,
+              'submission'
+            )
         }
         { autograded ? this.renderAutogradedFields() : null }
       </div>
@@ -354,7 +499,7 @@ class OnlineEditorCppView extends React.Component {
   }
 }
 
-OnlineEditorCppView.propTypes = propTypes;
-OnlineEditorCppView.contextTypes = contextTypes;
+OnlineEditorJavaView.propTypes = propTypes;
+OnlineEditorJavaView.contextTypes = contextTypes;
 
-export default injectIntl(OnlineEditorCppView);
+export default injectIntl(OnlineEditorJavaView);
